renamed IPv4Network class file
